requirements.txt
twikit
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-dotenv
chromadb
langchain
langchain-groq
langchain-community
langgraph
beautifulsoup4
requests
aiohttp
groq
aiosqlite
playwright
feedparser
duckduckgo-search
transformers
twikit
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-dotenv
chromadb
langchain
langchain-groq
langchain-community
langgraph
beautifulsoup4
requests
aiohttp
groq
aiosqlite
playwright
feedparser
duckduckgo-search
transformers
torch
pillow
einops
yt-dlp
openai-whisper
ffmpeg-python
aiokafka
scikit-learn
googlesearch-python
xmltodict
torchvision